+++ nodalia-cards.manifest.js
@@ -1,11 +1,11 @@
 export default {
-  "pkgVersion": "2.2.7",
-  "contentSha256_12": "67262e83d56b",
+  "pkgVersion": "2.2.8",
+  "contentSha256_12": "cd9865b537dc",
   "file": "nodalia-cards.js",
   "loaderFile": "nodalia-cards.js",
   "hacsFile": "nodalia-cards.js",
   "editorSha256_12": "9722b9c0fc8f"
 };
-export const pkgVersion = "2.2.7";
-export const contentSha256_12 = "67262e83d56b";
+export const pkgVersion = "2.2.8";
+export const contentSha256_12 = "cd9865b537dc";
 export const file = "nodalia-cards.js";